+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Improved the startup time of `import civis` with a 5x speed boost. (#490)
 - The downloaded API spec due to the `civis.APIClient` instantiation is now
   a time-to-live cache in memory (15 minutes for interactive Python, or 24 hours in scripts). (#491)
+- Polling at `PollableResult` (and consequently its subclasses as well: `CivisFuture`,
+  `ContainerFuture`, and `ModelFuture`) now defaults to geometrically increased polling
+  intervals and is capped at 15 seconds. Short-running jobs' `future.result()` can now
+  return faster and longer-running jobs have a capped polling interval of 15 seconds. (#492)
 
 ### Deprecated
 ### Removed